+++ lib/src/colors/brown.dart
@@ -1,0 +1,101 @@
+import 'pumpkin_orange.dart';
+
+/// Brown with 95% Lightness.
+///
+/// Hex Code: #F9F1EC
+const brown50 = pumpkinOrangeJ50;
+
+/// Brown with 90% Lightness.
+///
+/// Hex Code: #F2E3D9
+const brown100 = pumpkinOrangeJ100;
+
+/// Brown with 85% Lightness.
+///
+/// Hex Code: #ECD5C6
+const brown150 = pumpkinOrangeJ150;
+
+/// Brown with 80% Lightness.
+///
+/// Hex Code: #E6C7B3
+const brown200 = pumpkinOrangeJ200;
+
+/// Brown with 75% Lightness.
+///
+/// Hex Code: #DFB99F
+const brown250 = pumpkinOrangeJ250;
+
+/// Brown with 70% Lightness.
+///
+/// Hex Code: #D9AB8C
+const brown300 = pumpkinOrangeJ300;
+
+/// Brown with 65% Lightness.
+///
+/// Hex Code: #D29D79
+const brown350 = pumpkinOrangeJ350;
+
+/// Brown with 60% Lightness.
+///
+/// Hex Code: #CC8F66
+const brown400 = pumpkinOrangeJ400;
+
+/// Brown with 55% Lightness.
+///
+/// Hex Code: #C68153
+const brown450 = pumpkinOrangeJ450;
+
+/// Brown with 50% Lightness.
+///
+/// Hex Code: #BF7340
+const brown500 = pumpkinOrangeJ500;
+
+/// Shorthand for [brown500], since it is the base color.
+///
+/// Hex Code: #BF7340
+const brown = brown500;
+
+/// Brown with 45% Lightness.
+///
+/// Hex Code: #AC6739
+const brown550 = pumpkinOrangeJ550;
+
+/// Brown with 40% Lightness.
+///
+/// Hex Code: #995C33
+const brown600 = pumpkinOrangeJ600;
+
+/// Brown with 35% Lightness.
+///
+/// Hex Code: #86502D
+const brown650 = pumpkinOrangeJ650;
+
+/// Brown with 30% Lightness.
+///
+/// Hex Code: #734526
+const brown700 = pumpkinOrangeJ700;
+
+/// Brown with 25% Lightness.
+///
+/// Hex Code: #603920
+const brown750 = pumpkinOrangeJ750;
+
+/// Brown with 20% Lightness.
+///
+/// Hex Code: #4D2E1A
+const brown800 = pumpkinOrangeJ800;
+
+/// Brown with 15% Lightness.
+///
+/// Hex Code: #392213
+const brown850 = pumpkinOrangeJ850;
+
+/// Brown with 10% Lightness.
+///
+/// Hex Code: #26170D
+const brown900 = pumpkinOrangeJ900;
+
+/// Brown with 5% Lightness.
+///
+/// Hex Code: #130B06
+const brown950 = pumpkinOrangeJ950;