+++ lib/src/colors/grey.dart
@@ -1,115 +1,101 @@
 import 'dart:ui';
 
-/// Grey with 95% Lightness. Hex Code: #F2F2F2.
+/// Grey with 95% Lightness.
+///
+/// Hex Code: #F2F2F2.
 const grey50 = Color(0xffF2F2F2);
 
-/// Grey with 90% Lightness. Hex Code: #E6E6E6.
+/// Grey with 90% Lightness.
+///
+/// Hex Code: #E6E6E6.
 const grey100 = Color(0xffE6E6E6);
 
-/// Grey with 85% Lightness. Hex Code: #D9D9D9.
+/// Grey with 85% Lightness.
+///
+/// Hex Code: #D9D9D9.
 const grey150 = Color(0xffD9D9D9);
 
-/// Grey with 80% Lightness. Hex Code: #CCCCCC.
+/// Grey with 80% Lightness.
+///
+/// Hex Code: #CCCCCC.
 const grey200 = Color(0xffCCCCCC);
 
-/// Grey with 75% Lightness. Hex Code: #BFBFBF.
+/// Grey with 75% Lightness.
+///
+/// Hex Code: #BFBFBF.
 const grey250 = Color(0xffBFBFBF);
 
-/// Grey with 70% Lightness. Hex Code: #B3B3B3.
+/// Grey with 70% Lightness.
+///
+/// Hex Code: #B3B3B3.
 const grey300 = Color(0xffB3B3B3);
 
-/// Grey with 65% Lightness. Hex Code: #A6A6A6.
+/// Grey with 65% Lightness.
+///
+/// Hex Code: #A6A6A6.
 const grey350 = Color(0xffA6A6A6);
 
-/// Grey with 60% Lightness. Hex Code: #999999.
+/// Grey with 60% Lightness.
+///
+/// Hex Code: #999999.
 const grey400 = Color(0xff999999);
 
-/// Grey with 55% Lightness. Hex Code: #8C8C8C.
+/// Grey with 55% Lightness.
+///
+/// Hex Code: #8C8C8C.
 const grey450 = Color(0xff8C8C8C);
 
-/// Grey with 50% Lightness. Hex Code: #808080.
+/// Grey with 50% Lightness.
+///
+/// Hex Code: #808080.
 const grey500 = Color(0xff808080);
 
-/// Grey with 45% Lightness. Hex Code: #737373.
+/// Shorthand for [grey500], since it is the base color.
+///
+/// Hex Code: #808080.
+const grey = grey500;
+
+/// Grey with 45% Lightness.
+///
+/// Hex Code: #737373.
 const grey550 = Color(0xff737373);
 
-/// Grey with 40% Lightness. Hex Code: #666666.
+/// Grey with 40% Lightness.
+///
+/// Hex Code: #666666.
 const grey600 = Color(0xff666666);
 
-/// Grey with 35% Lightness. Hex Code: #595959.
+/// Grey with 35% Lightness.
+///
+/// Hex Code: #595959.
 const grey650 = Color(0xff595959);
 
-/// Grey with 30% Lightness. Hex Code: #4D4D4D.
+/// Grey with 30% Lightness.
+///
+/// Hex Code: #4D4D4D.
 const grey700 = Color(0xff4D4D4D);
 
-/// Grey with 25% Lightness. Hex Code: #404040.
+/// Grey with 25% Lightness.
+///
+/// Hex Code: #404040.
 const grey750 = Color(0xff404040);
 
-/// Grey with 20% Lightness. Hex Code: #333333.
+/// Grey with 20% Lightness.
+///
+/// Hex Code: #333333.
 const grey800 = Color(0xff333333);
 
-/// Grey with 15% Lightness. Hex Code: #262626.
+/// Grey with 15% Lightness.
+///
+/// Hex Code: #262626.
 const grey850 = Color(0xff262626);
 
-/// Grey with 10% Lightness. Hex Code: #1A1A1A.
+/// Grey with 10% Lightness.
+///
+/// Hex Code: #1A1A1A.
 const grey900 = Color(0xff1A1A1A);
 
-/// Grey with 5% Lightness. Hex Code: #0D0D0D.
+/// Grey with 5% Lightness.
+///
+/// Hex Code: #0D0D0D.
 const grey950 = Color(0xff0D0D0D);
-
-/// Gray with 95% Lightness. Hex Code: #F2F2F2.
-const gray50 = Color(0xffF2F2F2);
-
-/// Gray with 90% Lightness. Hex Code: #E6E6E6.
-const gray100 = Color(0xffE6E6E6);
-
-/// Gray with 85% Lightness. Hex Code: #D9D9D9.
-const gray150 = Color(0xffD9D9D9);
-
-/// Gray with 80% Lightness. Hex Code: #CCCCCC.
-const gray200 = Color(0xffCCCCCC);
-
-/// Gray with 75% Lightness. Hex Code: #BFBFBF.
-const gray250 = Color(0xffBFBFBF);
-
-/// Gray with 70% Lightness. Hex Code: #B3B3B3.
-const gray300 = Color(0xffB3B3B3);
-
-/// Gray with 65% Lightness. Hex Code: #A6A6A6.
-const gray350 = Color(0xffA6A6A6);
-
-/// Gray with 60% Lightness. Hex Code: #999999.
-const gray400 = Color(0xff999999);
-
-/// Gray with 55% Lightness. Hex Code: #8C8C8C.
-const gray450 = Color(0xff8C8C8C);
-
-/// Gray with 50% Lightness. Hex Code: #808080.
-const gray500 = Color(0xff808080);
-
-/// Gray with 45% Lightness. Hex Code: #737373.
-const gray550 = Color(0xff737373);
-
-/// Gray with 40% Lightness. Hex Code: #666666.
-const gray600 = Color(0xff666666);
-
-/// Gray with 35% Lightness. Hex Code: #595959.
-const gray650 = Color(0xff595959);
-
-/// Gray with 30% Lightness. Hex Code: #4D4D4D.
-const gray700 = Color(0xff4D4D4D);
-
-/// Gray with 25% Lightness. Hex Code: #404040.
-const gray750 = Color(0xff404040);
-
-/// Gray with 20% Lightness. Hex Code: #333333.
-const gray800 = Color(0xff333333);
-
-/// Gray with 15% Lightness. Hex Code: #262626.
-const gray850 = Color(0xff262626);
-
-/// Gray with 10% Lightness. Hex Code: #1A1A1A.
-const gray900 = Color(0xff1A1A1A);
-
-/// Gray with 5% Lightness. Hex Code: #0D0D0D.
-const gray950 = Color(0xff0D0D0D);