+++ lib/src/colors/slate.dart
@@ -1,0 +1,101 @@
+import 'cornflower_blue.dart';
+
+/// Slate with 95% Lightness.
+///
+/// Hex Code: #F1F2F4
+const slate50 = cornflowerBlueR50;
+
+/// Slate with 90% Lightness.
+///
+/// Hex Code: #E3E5E8
+const slate100 = cornflowerBlueR100;
+
+/// Slate with 85% Lightness.
+///
+/// Hex Code: #D5D8DD
+const slate150 = cornflowerBlueR150;
+
+/// Slate with 80% Lightness.
+///
+/// Hex Code: #C7CBD1
+const slate200 = cornflowerBlueR200;
+
+/// Slate with 75% Lightness.
+///
+/// Hex Code: #B9BEC6
+const slate250 = cornflowerBlueR250;
+
+/// Slate with 70% Lightness.
+///
+/// Hex Code: #ABB1BA
+const slate300 = cornflowerBlueR300;
+
+/// Slate with 65% Lightness.
+///
+/// Hex Code: #9DA4AF
+const slate350 = cornflowerBlueR350;
+
+/// Slate with 60% Lightness.
+///
+/// Hex Code: #8F97A3
+const slate400 = cornflowerBlueR400;
+
+/// Slate with 55% Lightness.
+///
+/// Hex Code: #818A98
+const slate450 = cornflowerBlueR450;
+
+/// Slate with 50% Lightness.
+///
+/// Hex Code: #737D8C
+const slate500 = cornflowerBlueR500;
+
+/// Shorthand for [slate500], since it is the base color.
+///
+/// Hex Code: #737D8C
+const slate = slate500;
+
+/// Slate with 45% Lightness.
+///
+/// Hex Code: #67707E
+const slate550 = cornflowerBlueR550;
+
+/// Slate with 40% Lightness.
+///
+/// Hex Code: #5C6470
+const slate600 = cornflowerBlueR600;
+
+/// Slate with 35% Lightness.
+///
+/// Hex Code: #505762
+const slate650 = cornflowerBlueR650;
+
+/// Slate with 30% Lightness.
+///
+/// Hex Code: #454B54
+const slate700 = cornflowerBlueR700;
+
+/// Slate with 25% Lightness.
+///
+/// Hex Code: #393E46
+const slate750 = cornflowerBlueR750;
+
+/// Slate with 20% Lightness.
+///
+/// Hex Code: #2E3238
+const slate800 = cornflowerBlueR800;
+
+/// Slate with 15% Lightness.
+///
+/// Hex Code: #22252A
+const slate850 = cornflowerBlueR850;
+
+/// Slate with 10% Lightness.
+///
+/// Hex Code: #17191C
+const slate900 = cornflowerBlueR900;
+
+/// Slate with 5% Lightness.
+///
+/// Hex Code: #0B0C0E
+const slate950 = cornflowerBlueR950;